Implement `MemoryPool` and `MemoryPoolZone` in C to manage zone memory.
In current unsafe zone, the zone memory is in the heap, managed by the
Scala-implemented `MemoryPoolZone` in nativelib. This commit implement
`MemoryPool` which manage a new memory area other than heap and stack, and
`MemoryPoolZone` which implements the zone operations based on the
memory pool. This enables the zone operations to be backend operations
as GC operations.